Add files via upload
We define 8000 images with numpy, using random normal noise, in a way that some of them are visibly bright and others visibly dark in relation to each other. Then, we normalize the data and feed it into the neural network. We save the model and create more random images and test the model's predictions in this set, plotting the results with matplotlib.